pyproject.toml
[tool.poetry]
name = "afetchx"
version = "1.1.0"
description = ""
authors = ["DJ Stomp <85457381+DJStompZone@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
tqdm = "^4.66.6"
httpx = "^0.27.2"
argparse = "^1.4.0"
aiofiles = "^24.1.0"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
